+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Slowly increase the delay in keepTrying-calls to reduce resource usage.
 - Fix visiblity detection of Clarity modals without title.
 - Fix and optimize selectByIndex
+- Add class utilities to SeleniumUtils
 
 ## selenium-components-0.4.3